[ ADD,EDIT-layout ] 공통 레이아웃인 navBar컴포넌트를 구현하라 - 각 페이지에 공통적으로 있는 navbar를 layout으로 빼어 구현하였다. - 해당 navbar가 있는 코드는 삭제하였다. - 로그인/ 회원가입 페이지에도 navbar가 생겼다..그건 좀 문제다 ㅋㅋ 해결해야함
