+++ app/components/pageLayout.module.css
@@ -1,0 +1,18 @@
+.nav {
+  width: 500px;
+  height: 60px;
+  margin: 0 auto;
+  padding: 0 16px;
+  box-sizing: border-box;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: #0e4f3d;
+  color: white;
+}
+.nav .navLogo {
+  color: white;
+  font-size: 25px;
+  font-weight: 400;
+  cursor: pointer;
+}
