+++ app/mypage/mypage.module.css
@@ -5,27 +5,6 @@
   box-shadow: 1px 1px 5px #eee;
   background-color: white;
 }
-
-.nav {
-  width: 100%;
-  height: 60px;
-  padding: 0 16px;
-  box-sizing: border-box;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #0e4f3d;
-}
-.nav .navLogo {
-  color: white;
-  font-size: 25px;
-  font-weight: 400;
-  cursor: pointer;
-}
-.nav .loginBtn {
-  color: white;
-  cursor: pointer;
-}
 .userInfoBox{
   padding: 16px;
   text-align: center;
