+++ app/postPage/postPage.module.css
@@ -6,23 +6,6 @@
   border: 1px solid #e6e6e6;
   background-color: white;
 }
-.nav {
-  width: 100%;
-  height: 60px;
-  padding: 0 16px;
-  box-sizing: border-box;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #0e4f3d;
-  color: white;
-}
-.nav .navLogo {
-  color: white;
-  font-size: 25px;
-  font-weight: 400;
-  cursor: pointer;
-}
 
 .postBox {
 }
@@ -83,14 +66,14 @@
   align-items: center;
   background-color: #f4f4f4;
 }
-.postBox .postImageBox .labelBox{
+.postBox .postImageBox .labelBox {
   text-align: center;
   color: gray;
 }
-.postBox .postImageBox .labelBox label{
+.postBox .postImageBox .labelBox label {
   font-size: 18px;
   cursor: pointer;
 }
-.postBox .postImageBox input[type=file]{
+.postBox .postImageBox input[type="file"] {
   display: none;
 }
